% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{nat_palette}
\alias{nat_palette}
\title{A 'natural' colour palette generator}
\usage{
nat_palette(name, n)
}
\arguments{
\item{name}{The desired palette. One of:
\itemize{
  \item \code{Acripeza}: The mountain katydid \emph{Acripeza reticulata}.
  \item \code{Aix}: The mandarin duck \emph{Aix galericulata}.
  \item \code{Chrysiridia}: The Madagascan sunset moth \emph{Chrysiridia rhipheus}.
  \item \code{Coracias}: The lilac-breasted roller \emph{Coracias caudatus}.
  \item \code{Furcifur}: The panther chameleon \emph{Furcifer pardalis}.
  \item \code{Delias}: The Phillippine jezabel \emph{Delias henningia}.
  \item \code{Maratus}: The peacock spider \emph{Maratus volans}.
  \item \code{Synchiropus}: The mandarinfish \emph{Synchiropus splendidus}.
  \item \code{Trichoglossus}: The rainbow lorikeet \emph{Trichoglossus haematodus}.
  \item \code{Tulipa}: Tulips \emph{Tulipa gesneriana}.
  }}

\item{n}{Number of colours. If omitted, uses all colours.}
}
\value{
A vector of colours.
}
\description{
Some fairly impractical colour palettes inspired by nature.
}
\examples{
nat_palette('Maratus')
nat_palette('Tulipa')

}
\keyword{colours}
