% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Palette list}
\format{An object of class \code{list} of length 10.}
\usage{
palettes
}
\description{
Colour list. Use \code{\link{nat_palette}} to build palettes.
}
\keyword{datasets}

